Split labels out of `prometheus.rs` into its own file
